move: auth와 member로 나누기 위해 controller, service, test 모두 auth 폴더 생성후 로직 분리
