+++ README.md
@@ -55,13 +55,13 @@
 
 ### 나만의 HTTP RULE
 
-| HTTP Method | 사용상황                          | 반환(상태코드) |
-|-------------|-------------------------------|----------|
-| GET         | 리소스의 조회                       | 200      | 
-| POST        | 새로운 리소스 생성                    | 201      |
-| PUT         | 리소스의 전체 업데이트 또는 ID를 통한 리소스 생성 | 204      |
-| PATCH       | 리소스의 일부분(일부 필드) 업데이트          | 204      |
-| DELETE      | 리소스의 삭제                       | 204      |
+| HTTP Method | 사용상황                           | 반환(상태코드) |
+|-------------|--------------------------------|----------|
+| GET         | 리소스의 조회                        | 200      | 
+| POST        | 새로운 리소스 생성                     | 201      |
+| PUT         | 리소스의 전체 업데이트 또는 ID를 통한 리소스 생성  | 204      |
+| PATCH       | 리소스의 일부분(일부 필드) 업데이트           | 204      |
+| DELETE      | 리소스의 삭제                        | 204      |
 
 ### 나만의 계층 RULE
